backend/requirements-dev.txt
# Include all production dependencies
-r requirements.txt

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
coverage==7.4.1
httpx==0.27.0
requests==2.31.0
iniconfig==2.0.0

# Development tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Debug tools
ipython==8.21.0
ipdb==0.13.13

# Code Quality
flake8-docstrings==1.7.0  # Docstring checking
flake8-quotes==3.4.0  # Quote style checking
pylint==3.0.3  # Code analysis
bandit==1.7.7  # Security linting
safety==2.3.5  # Dependency security checking

# Development Tools
watchdog==3.0.0  # File system events
pre-commit==3.6.0  # Git hooks
pre-commit-hooks==4.5.0  # Additional pre-commit hooks
blacken-docs==1.13.0  # Format docstrings
isort[pyproject]==5.13.2  # Import sorting with pyproject.toml support

# Documentation
mkdocstrings==0.24.0  # Auto-documentation
mkdocstrings-python==1.7.0  # Python handler for mkdocstrings
mkdocs-autorefs==0.4.0  # Automatic cross-references
mkdocs-minify-plugin==1.7.0  # Minify HTML/CSS/JS
mkdocs-redirects==1.2.1  # Handle redirects
mkdocs-swagger-ui-tag==0.1.8  # Swagger UI integration

# Debugging & Profiling
debugpy==1.8.0  # Remote debugging
memory-profiler==0.61.0  # Memory profiling
line-profiler==4.1.1  # Line-by-line profiling
py-spy==0.3.14  # Sampling profiler
objgraph==3.6.0  # Object graph visualization

# Database Tools
alembic==1.13.1  # Database migrations
sqlalchemy-utils==0.41.1  # SQLAlchemy utilities
mysqlclient==2.2.1  # MySQL client
pymysql==1.1.0  # Pure Python MySQL client

# API Development
fastapi-cli==0.1.0  # FastAPI CLI tools
fastapi-debug-toolbar==0.3.0  # Debug toolbar
fastapi-sqlalchemy==0.4.0  # SQLAlchemy integration
fastapi-cache2==0.2.1  # Caching support
fastapi-limiter==0.1.5  # Rate limiting
fastapi-pagination==0.12.14  # Pagination support

# Monitoring & Logging
prometheus-client==0.19.0  # Prometheus metrics
python-json-logger==2.0.7  # JSON logging
structlog==24.1.0  # Structured logging
sentry-sdk==1.39.1  # Error tracking
opentelemetry-api==1.21.0  # OpenTelemetry API
opentelemetry-sdk==1.21.0  # OpenTelemetry SDK
opentelemetry-instrumentation-fastapi==0.42b0  # FastAPI instrumentation

# Task Queue Development
celery==5.3.6  # Task queue
flower==2.0.1  # Celery monitoring
redis==5.0.1  # Redis client

# Email Development
aiosmtplib==3.0.1  # Async SMTP
jinja2==3.1.3  # Template engine
email-validator==2.1.0.post1  # Email validation

# Utilities
python-dateutil==2.8.2  # Date utilities
pytz==2024.1  # Timezone support
tqdm==4.66.1  # Progress bars
tenacity==8.2.3  # Retry logic
python-dotenv==1.0.1  # Environment variables
click==8.1.7  # Command line interface
rich==13.7.0  # Terminal formatting 